jsencoding: fix incorrect conversion

+++ src/html/jsencoding.nim
@@ -96,7 +96,7 @@ proc encode(this: JSTextEncoder; input = ""): JSArrayBufferView {.jsfunc.} =
     t: JS_TYPED_ARRAY_UINT8,
     abuf: JSArrayBuffer(p: p, len: csize_t(input.len), dealloc: deallocWrap),
     offset: 0,
-    len: int64(input.len)
+    len: input.len
   )
 
 #TODO encodeInto
